Move keyboard/mouse bindings to nested fields
Having elements at the root level of the config forces them at the top
of the file. For bindings which are usually just a big list this makes
things really ugly. To work around that we just move the keyboard and
mouse bindings to a separate struct.
